travistooling: add code for unpacking secrets
