--- travistooling/travisenv.py
@@ -1,12 +1,0 @@
-# -*- encoding: utf-8
-
-import os
-
-
-def branch_name():
-    """Return the name of the branch under test."""
-    # See https://graysonkoonce.com/getting-the-current-branch-name-during-a-pull-request-in-travis-ci/
-    if os.environ['TRAVIS_PULL_REQUEST'] == 'false':
-        return os.environ['TRAVIS_BRANCH']
-    else:
-        return os.environ['TRAVIS_PULL_REQUEST_BRANCH']
